+++ powerjob-server/src/main/java/com/github/kfcfans/powerjob/server/OhMyApplication.java
@@ -1,6 +1,6 @@
 package com.github.kfcfans.powerjob.server;
 
-import com.github.kfcfans.powerjob.server.akka.OhMyServer;
+import com.github.kfcfans.powerjob.server.transport.akka.OhMyServer;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
